Conflicts:
	src/3rdparty/javascriptcore/JavaScriptCore/ChangeLog
	src/3rdparty/javascriptcore/JavaScriptCore/JavaScriptCore.pri
	src/3rdparty/javascriptcore/JavaScriptCore/interpreter/RegisterFile.cpp
	src/3rdparty/javascriptcore/JavaScriptCore/interpreter/RegisterFile.h
	src/3rdparty/javascriptcore/JavaScriptCore/runtime/JSGlobalData.cpp
	src/3rdparty/javascriptcore/JavaScriptCore/runtime/JSGlobalData.h
	src/3rdparty/javascriptcore/JavaScriptCore/runtime/JSGlobalObject.h
	src/3rdparty/javascriptcore/JavaScriptCore/runtime/MathObject.cpp
	src/3rdparty/javascriptcore/JavaScriptCore/wtf/RandomNumber.h
	src/3rdparty/javascriptcore/VERSION
	src/script/api/qscriptcontext.cpp
	src/script/api/qscriptengine.cpp
	src/script/script.pro

+++ bin/createpackage.pl
@@ -139,11 +139,18 @@ unless (GetOptions('i|install' => \$install,
     Usage();
 }
 
+my $epocroot = $ENV{EPOCROOT};
+$epocroot =~ s,[\\/]$,,x;
+
 my $certfilepath = abs_path(dirname($certfile));
 
 # Read params to variables
 my $templatepkg = $ARGV[0];
 my $targetplatform = lc $ARGV[1];
+
+if ($targetplatform eq "") {
+    $targetplatform = "-";
+}
 
 my @tmpvalues = split('-', $targetplatform);
 my $target;
@@ -176,11 +183,11 @@ $passphrase = $ARGV[4] or $passphrase = "";
 my $pkgoutputbasename = $templatepkg;
 my $preservePkgOutput = "";
 $pkgoutputbasename =~ s/_template/_$targetplatform/g;
+$pkgoutputbasename =~ s/_installer\.pkg/_installer___temp\.pkg/g;
 if ($pkgoutputbasename eq $templatepkg) {
     $preservePkgOutput = "1";
 }
 $pkgoutputbasename =~ s/\.pkg//g;
-$pkgoutputbasename = $pkgoutputbasename;
 
 # Store output file names to variables
 my $pkgoutput = $pkgoutputbasename.".pkg";
@@ -188,6 +195,7 @@ my $sisoutputbasename;
 if ($signed_sis_name eq "") {
     $sisoutputbasename = $pkgoutputbasename;
     $sisoutputbasename =~ s/_$targetplatform//g;
+    $sisoutputbasename =~ s/_installer___temp/_installer/g;
     $signed_sis_name = $sisoutputbasename.".sis";
 } else {
     $sisoutputbasename = $signed_sis_name;
@@ -196,6 +204,16 @@ if ($signed_sis_name eq "") {
     } else {
         $signed_sis_name = $signed_sis_name.".sis";
     }
+}
+
+my $installer_unsigned_app_sis_name = "";
+my $installer_app_sis_name = "";
+
+if ($templatepkg =~ m/_installer\.pkg$/i && $onlyUnsigned) {
+    $installer_unsigned_app_sis_name = $templatepkg;
+    $installer_unsigned_app_sis_name =~ s/_installer.pkg$/_unsigned.sis/i;
+    $installer_app_sis_name = $installer_unsigned_app_sis_name;
+    $installer_app_sis_name =~ s/_unsigned.sis$/.sis/;
 }
 
 my $unsigned_sis_name = $sisoutputbasename."_unsigned.sis";
@@ -209,21 +227,21 @@ my $certpath = File::Spec->catdir($scriptpath, File::Spec->updir(), "src/s60inst
 
 # Check some pre-conditions and print error messages if needed.
 unless (length($templatepkg)) {
-    print "\nError: Template PKG filename is not defined!\n";
+    print "\nERROR: Template PKG filename is not defined!\n";
     Usage();
 }
 
 # Check template exist
 stat($templatepkg);
 unless( -e _ ) {
-    print "\nError: Package description file '$templatepkg' does not exist!\n";
+    print "\nERROR: Package description file '$templatepkg' does not exist!\n";
     Usage();
 }
 
 # Check certifcate preconditions and set default certificate variables if needed
 if (length($certificate)) {
     unless(length($key)) {
-        print "\nError: Custom certificate key file parameter missing.!\n";
+        print "\nERROR: Custom certificate key file parameter missing.!\n";
         Usage();
     }
 } else {
@@ -258,7 +276,7 @@ if (length($certfile)) {
 
         # Do some validation
         unless(scalar(@certinfo) >= 2 && scalar(@certinfo) <= 3 && length($certinfo[0]) && length($certinfo[1]) ) {
-            print "\nError: $certfile line '$_' does not contain valid information!\n";
+            print "\nERROR: $certfile line '$_' does not contain valid information!\n";
             Usage();
         }
 
@@ -268,7 +286,9 @@ if (length($certfile)) {
 
 # Remove any existing .sis packages
 unlink $unsigned_sis_name;
-unlink $signed_sis_name;
+if (!$onlyUnsigned) {
+    unlink $signed_sis_name;
+}
 if (!$preservePkgOutput) {
     unlink $pkgoutput;
 }
@@ -277,14 +297,14 @@ if (!$preservePkgOutput) {
 
 local $/;
 # read template file
-open( TEMPLATE, $templatepkg) or die "Error '$templatepkg': $!\n";
+open( TEMPLATE, $templatepkg) or die "ERROR: '$templatepkg': $!";
 $_=<TEMPLATE>;
 close (TEMPLATE);
 
 # If the pkg file does not contain macros, there is no need for platform or target.
 if (m/\$\(PLATFORM\)/) {
     unless (length($platform) && length($target)) {
-        print "\nError: Platform or target is not defined!\n";
+        print "\nERROR: Platform or target is not defined!\n";
         Usage();
     }
 }
@@ -293,8 +313,12 @@ if (m/\$\(PLATFORM\)/) {
 s/\$\(PLATFORM\)/$platform/gm;
 s/\$\(TARGET\)/$target/gm;
 
+if ($installer_unsigned_app_sis_name ne "") {
+    s/$installer_app_sis_name\"/$installer_unsigned_app_sis_name\"/;
+}
+
 #write the output
-open( OUTPUT, ">$pkgoutput" ) or die "Error '$pkgoutput' $!\n";
+open( OUTPUT, ">$pkgoutput" ) or die "ERROR: '$pkgoutput' $!";
 print OUTPUT $_;
 close OUTPUT;
 
@@ -303,12 +327,12 @@ if ($preprocessonly) {
 }
 
 if($stub) {
-    if(!($ENV{EPOCROOT})) { die("EPOCROOT must be set to create stub sis files"); }
-    my $systeminstall = "$ENV{EPOCROOT}epoc32/data/z/system/install";
+    if(!($epocroot)) { die("ERROR: EPOCROOT must be set to create stub sis files"); }
+    my $systeminstall = "$epocroot/epoc32/data/z/system/install";
     mkpath($systeminstall);
     my $stub_sis_name = $systeminstall."/".$stub_sis_name;
     # Create stub SIS.
-    system ("makesis -s $pkgoutput $stub_sis_name");
+    system ("$epocroot/epoc32/tools/makesis -s $pkgoutput $stub_sis_name");
 } else {
     if ($certtext eq "Self Signed"
         && !@certificates
@@ -316,12 +340,16 @@ if($stub) {
         && !$onlyUnsigned) {
         print("Auto-patching capabilities for self signed package.\n");
         my $patch_capabilities = File::Spec->catfile(dirname($0), "patch_capabilities");
-        system ("$patch_capabilities $pkgoutput");
+        system ("$patch_capabilities $pkgoutput") and die ("ERROR: Automatic patching failed");
     }
 
     # Create SIS.
     # The 'and' is because system uses 0 to indicate success.
-    system ("makesis $pkgoutput $unsigned_sis_name") and die ("makesis failed");
+    if($epocroot) {
+        system ("$epocroot/epoc32/tools/makesis $pkgoutput $unsigned_sis_name") and die ("ERROR: makesis failed");
+    } else {
+        system ("makesis $pkgoutput $unsigned_sis_name") and die ("ERROR: makesis failed");
+    }
     print("\n");
 
     my $targetInsert = "";
@@ -340,6 +368,7 @@ if($stub) {
         if (!$preservePkgOutput) {
             unlink $pkgoutput;
         }
+        print ("\n");
         exit;
     }
 
@@ -347,7 +376,7 @@ if($stub) {
     my $relcert = File::Spec->abs2rel($certificate);
     my $relkey = File::Spec->abs2rel($key);
     # The 'and' is because system uses 0 to indicate success.
-    system ("signsis $unsigned_sis_name $signed_sis_name $relcert $relkey $passphrase") and die ("signsis failed");
+    system ("signsis $unsigned_sis_name $signed_sis_name $relcert $relkey $passphrase") and die ("ERROR: signsis failed");
 
     # Check if creating signed SIS Succeeded
     stat($signed_sis_name);
@@ -381,6 +410,7 @@ if($stub) {
         # Lets leave the generated PKG for problem solving purposes
         print ("\nSIS creation failed!\n");
     }
+    print ("\n");
 }
 
 #end of file
